Update the sync API to use the new version of the API Framework.
