--- api/synchronous/api_core/GPUContainerGroup.yaml
@@ -1,34 +1,0 @@
-apiVersion: '2018-10-01'
-location: eastus
-name: animaldetectiongpu_cg
-properties:
-  containers:
-  - name: animaldetectiongpu
-    properties:
-      environmentVariables: []
-      image: yasiyu.azurecr.io/animal-detection:1-gpu
-      ports:
-      - port: 80
-      resources:
-        requests:
-          cpu: 1.0
-          gpu:
-            count: 1
-            sku: K80
-          memoryInGB: 5.0
-  ipAddress:
-    ports:
-    - port: 80
-      protocol: TCP
-    type: Public
-  osType: Linux
-  restartPolicy: OnFailure
-  shutdownGracePeriod: PT1M
-  timeToLive: PT1H
-  imageRegistryCredentials:
-  - server: yasiyu.azurecr.io
-    username: yasiyu
-    password: "insert password"
-tags:
-  null
-type: Microsoft.ContainerInstance/containerGroups
