+++ api/synchronous/api_core/supervisord.conf
@@ -3,7 +3,7 @@ nodaemon=true
 
 [program:gunicorn]
 directory=/app/animal_detection_classification_api/
-command=/usr/local/envs/ai4e_py_api/bin/gunicorn --workers=1 runserver:app -b 0.0.0.0:8024 --timeout 60
+command=gunicorn -b 0.0.0.0:1212 --workers 4 runserver:app
 stdout_logfile=/dev/stdout
 stdout_logfile_maxbytes=0
 stderr_logfile=/dev/stdout
