+++ data_management/annotations/annotation_constants.py
@@ -11,9 +11,11 @@ We use integer indices here; this is different than the API output .json file,
 where indices are string integers.
 """
 
-NUM_DETECTOR_CATEGORIES = 4  # this is for choosing colors, so ignoring the "empty" class
+NUM_DETECTOR_CATEGORIES = 3  # this is for choosing colors, so ignoring the "empty" class
 
-bbox_categories = [
+# This is the label mapping used for our incoming iMerit annotations
+# Only used to parse the incoming annotations. In our database, the string name is used to avoid confusion
+label_annotations_bbox_categories = [
     {'id': 0, 'name': 'empty'},
     {'id': 1, 'name': 'animal'},
     {'id': 2, 'name': 'person'},
@@ -21,9 +23,17 @@ bbox_categories = [
     {'id': 4, 'name': 'vehicle'}
 ]
 
-bbox_category_id_to_name = {}
-bbox_category_name_to_id = {}
+# MegaDetector outputs
+detector_bbox_categories = [
+    {'id': 0, 'name': 'empty'},
+    {'id': 1, 'name': 'animal'},
+    {'id': 2, 'name': 'person'},
+    {'id': 3, 'name': 'vehicle'}
+]
 
-for cat in bbox_categories:
-    bbox_category_id_to_name[cat['id']] = cat['name']
-    bbox_category_name_to_id[cat['name']] = cat['id']
+detector_bbox_category_id_to_name = {}
+detector_bbox_category_name_to_id = {}
+
+for cat in detector_bbox_categories:
+    detector_bbox_category_id_to_name[cat['id']] = cat['name']
+    detector_bbox_category_name_to_id[cat['name']] = cat['id']
