+++ environment.yml
@@ -32,4 +32,5 @@ dependencies:
   # doc: https://azuresdkdocs.blob.core.windows.net/$web/python/azure-cosmos/4.0.0b5/index.html
   - azure-cosmos==4.0.0b6  
   - azureml-sdk[tensorboard]==1.0.83
+  - requests_toolbelt  # required to call the sync API
